requirements.txt
opencv-python
pyserial
pyttsx3
numpy
